Revert "1st stage of copy/paste"
This reverts commit de18a997fe86ed22271710687bd80bcdb4aa6358.

+++ MALClient/Models/IAnimeData.cs
@@ -3,7 +3,7 @@
     public interface IAnimeData
     {
         int Id { get; }
-        float MyScore { get; set; }
+        int MyScore { get; set; }
         int MyEpisodes { get; set; }
         string Title { get; }
         int MyStatus { get; set; }
